+++ src/components/cart/constants.js
@@ -1,0 +1,5 @@
+import appConfig from 'app-config';
+
+export const ORDER_BTN_WIDTH = 100;
+export const CART_ITEM_WHITE_SPACE = 30;
+export const CART_ITEM_WIDTH = appConfig.device.width - ORDER_BTN_WIDTH - CART_ITEM_WHITE_SPACE;