+++ CHANGELOG.md
@@ -5,12 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
-- List scanned product.
 
-## [r11.3.10] - 2021-02-26
+## [r11.3.11] - 2021-03-03
+### Changed
+- CartItem also using text input to enter quantity.
+
+## [r11.3.10] - 2021-03-02
 ### Added
 - Logic for config key allow site sale out inventory.
+- List scanned product.
 - Check product code by qrcode.
 ### Fixed
 - Cart footer UI quantity.